Fix offset bug
Fixed a bug where the offset list was not of equal length to that of the token list. This resulted in the token strings having a wrong offset in the metadata. Added an Assertion to ensure these lists are of equal length. There is a test case for this in the 20 example test, which is commented out for now due to runtime.

- Small import correction as well